+++ bullet.hpp
@@ -7,7 +7,7 @@
 class Bullet {
 public:
   Bullet(float gamer_x, float gamer_y, float gamer_angle);
-  void Update(float time);
+  void Update(long double time);
   void Draw(sf::RenderWindow &window);
   sf::Rect<int> Get_rect();
 };
