+++ gamestate.hpp
@@ -5,21 +5,21 @@
 
 class GameState {
 public:
-  virtual void Update(float time) = 0;
+  virtual void Update(long double time) = 0;
   virtual void Draw(sf::RenderWindow &window) = 0;
   virtual ~GameState() = default;
 };
 
 class MenuState : public GameState {
 public:
-  void Update(float time);
+  void Update(long double time);
   void Draw(sf::RenderWindow &window);
 };
 
 class PlayingState : public GameState {
 public:
   PlayingState();
-  void Update(float time);
+  void Update(long double time);
   void Draw(sf::RenderWindow &window);
 };
